refactor: move controller logic into router file
